+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ## `add dynamic logger`
  - add dynamic logger by providing db collection so user can log to different collections
+
+ ### `add new log type`
+ - add logFatal to log types